+++ backend/canisters/user/impl/src/queries/public_profile.rs
@@ -6,14 +6,14 @@ use user_canister::public_profile::{Response::*, *};
 fn public_profile(_args: Args) -> Response {
     read_state(|state| {
         Success(PublicProfile {
-            username: state.data.username.value.clone(),
-            display_name: state.data.display_name.value.clone(),
-            avatar_id: state.data.avatar.id(),
-            profile_background_id: state.data.profile_background.id(),
-            bio: state.data.bio.value.clone(),
-            is_premium: state.data.phone_is_verified || state.data.storage_limit > 0,
-            phone_is_verified: state.data.phone_is_verified,
-            created: state.data.user_created,
+            username: state.data.user.username.value.clone(),
+            display_name: state.data.user.display_name.value.clone(),
+            avatar_id: state.data.user.avatar.id(),
+            profile_background_id: state.data.user.profile_background.id(),
+            bio: state.data.user.bio.value.clone(),
+            is_premium: state.data.user.phone_is_verified || state.data.user.storage_limit > 0,
+            phone_is_verified: state.data.user.phone_is_verified,
+            created: state.data.user.user_created,
         })
     })
 }