+++ backend/canisters/user/impl/src/updates/add_reaction.rs
@@ -19,9 +19,9 @@ fn add_reaction_impl(args: Args, state: &mut RuntimeState) -> OCResult {
         return Err(OCErrorCode::InvalidReaction.into());
     }
 
-    state.data.verify_not_suspended()?;
+    state.data.user.verify_not_suspended()?;
 
-    let chat = state.data.direct_chats.get_mut_or_err(&args.user_id.into())?;
+    let chat = state.data.user.direct_chats.get_mut_or_err(&args.user_id.into())?;
     let my_user_id = state.env.canister_id().into();
     let now = state.env.now();
 
@@ -50,9 +50,9 @@ fn add_reaction_impl(args: Args, state: &mut RuntimeState) -> OCResult {
             message_id: args.message_id,
             reaction: args.reaction,
             added: true,
-            username: state.data.username.value.clone(),
-            display_name: state.data.display_name.value.clone(),
-            user_avatar_id: state.data.avatar.id(),
+            username: state.data.user.username.value.clone(),
+            display_name: state.data.user.display_name.value.clone(),
+            user_avatar_id: state.data.user.avatar.id(),
         })),
     );