+++ backend/canisters/user/impl/src/updates/c2c_local_user_index.rs
@@ -31,23 +31,27 @@ fn process_event(event: LocalUserIndexEvent, state: &mut RuntimeState) {
 
     match event {
         LocalUserIndexEvent::UsernameChanged(ev) => {
-            state.data.username = Timestamped::new(ev.username, now);
+            state.data.user.username = Timestamped::new(ev.username, now);
         }
         LocalUserIndexEvent::DisplayNameChanged(ev) => {
-            state.data.display_name = Timestamped::new(ev.display_name, now);
+            state.data.user.display_name = Timestamped::new(ev.display_name, now);
             state.award_achievement_and_notify(Achievement::SetDisplayName, now);
         }
         LocalUserIndexEvent::PhoneNumberConfirmed(ev) => {
-            state.data.phone_is_verified = true;
-            state.data.storage_limit = ev.new_storage_limit;
+            state.data.user.phone_is_verified = true;
+            state.data.user.storage_limit = ev.new_storage_limit;
             openchat_bot::send_phone_number_confirmed_bot_message(&ev, state);
         }
         LocalUserIndexEvent::StorageUpgraded(ev) => {
-            state.data.storage_limit = ev.new_storage_limit;
+            state.data.user.storage_limit = ev.new_storage_limit;
             openchat_bot::send_storage_ugraded_bot_message(&ev, state);
         }
         LocalUserIndexEvent::ReferredUserRegistered(ev) => {
-            state.data.referrals.set_status(ev.user_id, ReferralStatus::Registered, now);
+            state
+                .data
+                .user
+                .referrals
+                .set_status(ev.user_id, ReferralStatus::Registered, now);
             openchat_bot::send_referred_user_joined_message(ev.user_id, ev.username, state);
         }
         LocalUserIndexEvent::UserSuspended(ev) => {
@@ -60,16 +64,18 @@ fn process_event(event: LocalUserIndexEvent, state: &mut RuntimeState) {
             // Check that the user didn't already leave the group before this event arrived
             if !state
                 .data
+                .user
                 .group_chats
                 .removed_since(ev.group_canister_timestamp)
                 .contains(&ev.chat_id)
             {
                 state
                     .data
+                    .user
                     .group_chats
                     .join(ev.chat_id, ev.local_user_index_canister_id, ev.latest_message_index, now);
 
-                state.data.hot_group_exclusions.remove(&ev.chat_id, now);
+                state.data.user.hot_group_exclusions.remove(&ev.chat_id, now);
                 state.award_achievement_and_notify(Achievement::JoinedGroup, now);
             }
         }
@@ -77,12 +83,14 @@ fn process_event(event: LocalUserIndexEvent, state: &mut RuntimeState) {
             // Check that the user didn't already leave the community before this event arrived
             if !state
                 .data
+                .user
                 .communities
                 .removed_since(ev.community_canister_timestamp)
                 .contains(&ev.community_id)
             {
                 let (community, _) = state
                     .data
+                    .user
                     .communities
                     .join(ev.community_id, ev.local_user_index_canister_id, now);
 
@@ -102,17 +110,17 @@ fn process_event(event: LocalUserIndexEvent, state: &mut RuntimeState) {
             }
         }
         LocalUserIndexEvent::DiamondMembershipPaymentReceived(ev) => {
-            let mut awarded = state.data.award_achievement(Achievement::UpgradedToDiamond, now);
+            let mut awarded = state.data.user.award_achievement(Achievement::UpgradedToDiamond, now);
 
             if matches!(ev.duration, DiamondMembershipPlanDuration::Lifetime) {
-                awarded |= state.data.award_achievement(Achievement::UpgradedToGoldDiamond, now);
+                awarded |= state.data.user.award_achievement(Achievement::UpgradedToGoldDiamond, now);
             }
 
             if awarded {
                 state.notify_user_index_of_chit(now);
             }
 
-            state.data.diamond_membership_expires_at = Some(ev.expires_at);
+            state.data.user.diamond_membership_expires_at = Some(ev.expires_at);
 
             if ev.send_bot_message {
                 openchat_bot::send_text_message(
@@ -123,7 +131,7 @@ fn process_event(event: LocalUserIndexEvent, state: &mut RuntimeState) {
                 );
             }
 
-            if let Some(referred_by) = state.data.referred_by {
+            if let Some(referred_by) = state.data.user.referred_by {
                 let status = if matches!(ev.duration, DiamondMembershipPlanDuration::Lifetime) {
                     ReferralStatus::LifetimeDiamond
                 } else {
@@ -134,9 +142,9 @@ fn process_event(event: LocalUserIndexEvent, state: &mut RuntimeState) {
         }
         LocalUserIndexEvent::NotifyUniquePersonProof(proof) => {
             state.award_achievement_and_notify(Achievement::ProvedUniquePersonhood, now);
-            state.data.unique_person_proof = Some(*proof);
+            state.data.user.unique_person_proof = Some(*proof);
 
-            if let Some(referred_by) = state.data.referred_by {
+            if let Some(referred_by) = state.data.user.referred_by {
                 state.push_user_canister_event(
                     referred_by,
                     UserCanisterEvent::SetReferralStatus(Box::new(ReferralStatus::UniquePerson)),
@@ -149,7 +157,7 @@ fn process_event(event: LocalUserIndexEvent, state: &mut RuntimeState) {
         LocalUserIndexEvent::ReinstateMissedDailyClaims(days) => state.reinstate_missed_daily_claims(days),
         LocalUserIndexEvent::BotRemoved(bot_id) => state.uninstall_bot(bot_id),
         LocalUserIndexEvent::BotUpdated(ev) => {
-            state.data.handle_bot_definition_updated(*ev, now);
+            state.data.user.handle_bot_definition_updated(*ev, now);
         }
     }
 }