+++ backend/canisters/user/impl/src/updates/pin_chat_v2.rs
@@ -18,16 +18,16 @@ fn pin_chat_impl(args: Args, state: &mut RuntimeState) -> Response {
 
     match args.chat {
         ChatInList::Direct(chat_id) => {
-            state.data.direct_chats.pin(chat_id, now);
+            state.data.user.direct_chats.pin(chat_id, now);
         }
         ChatInList::Group(chat_id) => {
-            state.data.group_chats.pin(chat_id, now);
+            state.data.user.group_chats.pin(chat_id, now);
         }
         ChatInList::Favourite(chat) => {
-            state.data.favourite_chats.pin(chat, now);
+            state.data.user.favourite_chats.pin(chat, now);
         }
         ChatInList::Community(community_id, channel_id) => {
-            if let Some(community) = state.data.communities.get_mut(&community_id) {
+            if let Some(community) = state.data.user.communities.get_mut(&community_id) {
                 community.pin(channel_id, now);
             } else {
                 return Response::Error(OCErrorCode::ChatNotFound.into());